inline comment (security: console_cmd option)
use of "x-terminal-emulator" for console_cmd would pose a security risk

'console' command
-----------------
Slim permits the user to type the special username 'console' to popup a terminal
emulator with the login program (sic) (?within). In the previous packages (<1.3.1)
the terminal command was set to x-terminal-emulator. That was ill-advised; it allowed
a terminal like gnome-terminal to be run, which has a feature to open new windows/tabs
which, instead of running login, resulted in access to a shell with root privileges.

Regarding this behavior, use of a terminal emulator like xterm is recommended.
The current SLiM package now (again) defaults to, and recommends, xterm.